mcp-gdelt/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-gdelt"
version = "1.0.0"
description = "Model Context Protocol server for the GDELT DOC 2.0 API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["mcp", "gdelt", "model-context-protocol", "news", "geopolitics"]

dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-gdelt = "mcp_gdelt.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gdelt"]

[tool.hatch.build.targets.sdist]
include = ["src/", "README.md", "pyproject.toml", ".env.example"]
